fix app.min.js: bring back codex's change
So it turned out that codex changed one line,
but then somehow the file got deleted - prob
because of its size. Bringing this line back now.

Also, tested this locally, and the plugin works.
It works both with ?p param and with perm
in config. The exported RDF, however, does
not capture the diagram correctly, but that's
irrelevant at this point because the plugin
works in principle, and we can revise later.