rdfexport: build, test, check in browser
seems to look well formatted and drawio
and rdf exports contain expected info.

upon reload of file the info is shown.

when i added multiple prefixes to cause
overflow of window, it added a vertical
scrollbar.

no acute issues therefore, as it seems.

there seems to be a few pixel misalign
between left edge of CSV path and
base URI textboxes, but that's fine.