run bun run test log and debug all
the results are about what I expected,
good!

also good to see that debug script
handles injection of config nicely
which was obvious because no new files
were originally generated.
i made this a bit nicer by supplying
pytest debug ontology for pytest
generated cases for maximum clarity.

also great to see that ontology_iri
override does not touch the default
prefix, which is exactly as is supposed to be
because prefix is generated by debug
script when it handles default injections,
so there is actually two points of injection
with one at default debug logic and another
one as illustrated in debug_cli_regression,
which is really nice and helpful!