run all tests and confirm ok
so there is much change in map
json and much unclassified,
but that's just because aa37
severely mocked v2 failed to
process now - expectedly.

Real changes only affected
knut-olborgs-forskningsnotater
and koronakommisjonen, and
in both there was a standalone
URI with rounded=1, so this
is the expected behavior.
I should probably just expose
this toggle in parses settings
so people could disable if they
want.

otherwise debug all and run test
both have pass/fail pattern as
expected, so I will merge this.

I considered limiting style denotes
literal check to only style check
without tokens and parent but
then decided to keep more
conservative because cases
where we might want to need this
probably would be simple rounded
boxes that need to be literalized.