checkpoint! 🍾 run test and debug all - clean logs
bun run test:log:macos; bun run debug:all:log:macos

there is perhaps still issues with non-main pytest
like debug tests or metabuilder tests, but did not
go into this yet.

also the fact that there is some duplicate and
overall "vibecoded" vibe code in overrides

The important thing is that the pipeline and fixtures
are processed as expected.