confirm that flowchart and class diagram have fixed curies - but include_label doesn not work
lack of propagation of parser config (although ontology_iri
does seem to propagate) is a known limitation pending a fix

good that curies seem fixed now; also tested in prod plugin
by changing Address to rdfs:Address, and this did get
recognized as an individual. However, changing shape to
rounded did not work for converting back to literal, which
is a bit sad and must be investigated.