fix all fixable tests - only real issues remain
so this works now except for the actual
rml serialisation issues that lead to mismatch
between map schema mapped and pipeline
mapped.

however, there is also that issue of mismatch
between fixture rml mapped and map schema
rml mapped for Authority, yet this is a different
issue (some Agent-CorporateBody) mismatch
yet needs to be dealt with.