upd rml fixtures and check csv
so the problem with test_map_schema_workflow.py
failing on authority went away after I took the
current .rml fixture.

also checked if csv in rml fixtures are actually
original and not preprocessed - correct.

so we are now completely good on map schema
workflow and can reliably use it as benchmark
for pipeline workflow.

so only pipeline workflow to fix now!