fix rmlmapper missing baseuri error
how for the purpose of tests, baseUri
must be specified, which is done.

In production, this should also work
as long as there is clear guidance in docs
that baseUri should NOT end with anything
and relative IRIs in mxCelss SHOULD start
with slash or hash like they do in Turtle.

Again, this only affects RMLmapping itself
and not the internal tests. So we should just
say that baseUri is interpreted in the sense
how rmlmapper interprets it, which makes
perfect sense.