MAJOR metabuilder: imports from legacy work now
This is a major bump because overrides can now
be used as standalone modules also, e.g., with
imports from legacy. which are not transfered
to metabuilt code. This is helpful to resolve
vicious typing imports (unsolvable using
TYPE_CHECKING) and overall, to run separately.

Note that TYPE_CHECKING may still be needed
for within-classes stuff which get type checked
within metabuilt code also.