fix the base uri issue
Of note, class-diagram-tweaked-compliant
scenario is now a good unit test for base URI,
prefix IRI and different combos and priority;
may as well be converted to a pytest test later.
ts_plugin and ts_pipeline can be inspected
and it is evident that this works as expected:
./data/debug/results/class-diagram-tweaked-compliant/

all comments are in code.

also in passing removed some
redundant code parts.

this seems to work - test all
logs attached. only preexisting
errors seems to remain.

also, the issue of incorrect minting
from curies is not addressed yet.