[wip] fix unresolved curies - but struggle with turtle base parser
so turtle serializes saves correct
turtle (with @base and default
ns set), which can be confirmed
from attached plugin output.

However, upon reparsing for
debug module, it fails to resolve
default prefixes appropriately -
they are now supported, by the
way, as can be seen from updated
class diagram tweaked compliant
fixture (:Address).

This is really more of an issue of
using rdflib to read graphs, so
should not in principle be a problem
for the plugin itself.