add support for layers
this is a Draw IO feature when
nodes can be added to named
layers within diagram; this way
layer is a mxCell with parent="0",
and cells within have its ID as
their parent vs "1" (1 is therefore
the default layer).

This looks like a natural analogy
for named graphs - placeholder
added for now but later can be
properly propagated to serializers.

I tested this and seems to work
with the new parent logic. No
special tests for layers yet; updated
older tests if any will come in
next commits.