ui only: reintegrate literal knobs from 046b50c
so i took the ui elements from here:

https://github.com/gbad-project/drawio/blob/046b50c6303940700e07b49a6927eb05649f1390/src/main/webapp/plugins/rdfexport/aicode/typescript_plugin/src/rdfexport.ts

and manually rewired them into
typescript codebase. looks nice,
testa all no fails (logs attached),
and on prod the elements are
visible and seem to work as expected
(e.g., default value loaded when not set,
new values added/changed if
changed in the ui, and empty array
if all entries removed; this persists
across reloads of file also).

so the next step will be to propagate
this to draw io parser

also, I decided not to do everything
in one go and will add knobs for
minting separately, after these are
properly propagated