restructure aicode docs
I contemplated that using 'docs' there
is misguiding because docs will be
top level. so i came up with this
convention - .chats for chat artifacts
and 'logs' and .reports for aicode
self-reports.

because AGENTS.md and
CHANGELOG.md were uses as context
for collaborative editing between
human and ai, i created .context for these.
it also makes sense because agent
can just be instructed to review
.context/ dir.