MAJOR: bump python to 3.14 -  test all ok
this was an easy decision to ensure the
maximum maintainability of this. because
python is only needed for development
here, I am not concerned about user
accessibility.

test all log attached - no failures.
luckily!

will commit updated uv.lock a bit
later because something is brewing
with sphinx in parallel