add initial sphinx setup
credit goes to gemini cli! i guided it
granularly and it did a good job

i launched this in my browser, and
there is still much to improve, but
we have what looks like a good
initial stub of docs to work from.

this adds some sphinx dev deps ofc