build_docs.py and bun run build:docs
this now copies all relevant files to docs/
and preprocesses them, which gives us
great control over this.

i tried to view in the browser, and looks
fine - at least files look viewable and none
seem to be missing.

issues:

- some links are still broken bc need to be
replaced with github URLs in source files to work,
or implemented in build_docs.py to replace
- not very informative titles for some pages,
like just readme
- gfm to sphinx alert syntax replacement not working