debug: propagate literal definitions from config
lol/kek from tbl.yml seem to be propagated to invokeDrawioParser now.

note that by default plugin won't pass
the default literal defs (i.e., when they
are not explicitly set in debug yml), which
may be a bug but really should be fine
because we can always set everything
directly in debug yml, and further draw io
pipeline in python should also read from
default yml if any are missing.

though here, if literal_definitions are not
set in debug config, debug passes empty
array to python which is definitely incorrect
behavior because this should be null.
need to fix.